drawToTexture can return RenderTexture2D

+++ src/layaAir/laya/display/Sprite.ts
@@ -24,6 +24,7 @@ import { Node } from "./Node";
 import { SpriteConst } from "./SpriteConst";
 import { Stage } from "./Stage";
 import { URL } from "../net/URL";
+import { RenderTexture2D } from "../resource/RenderTexture2D";
 
 
 /**在显示对象上按下后调度。
@@ -1188,8 +1189,8 @@ export class Sprite extends Node {
      * @param offsetX 
      * @param offsetY 
      */
-    drawToTexture(canvasWidth: number, canvasHeight: number, offsetX: number, offsetY: number): Texture {
-        return Sprite.drawToTexture(this, this._renderType, canvasWidth, canvasHeight, offsetX, offsetY);
+    drawToTexture(canvasWidth: number, canvasHeight: number, offsetX: number, offsetY: number, rt:RenderTexture2D|null=null): Texture|RenderTexture2D {
+        return Sprite.drawToTexture(this, this._renderType, canvasWidth, canvasHeight, offsetX, offsetY,rt);
     }
 
     /**
@@ -1247,29 +1248,40 @@ export class Sprite extends Node {
         return canv;
     }
 
+	static drawtocanvCtx:Context;
     /**
      * @private 
      * 
      */
-    static drawToTexture(sprite: Sprite, _renderType: number, canvasWidth: number, canvasHeight: number, offsetX: number, offsetY: number): Texture {
+    static drawToTexture(sprite: Sprite, _renderType: number, canvasWidth: number, canvasHeight: number, offsetX: number, offsetY: number, rt:RenderTexture2D|null=null): Texture|RenderTexture2D {
+		if(!Sprite.drawtocanvCtx){
+			Sprite.drawtocanvCtx = new Context();
+		}
         offsetX -= sprite.x;
         offsetY -= sprite.y;
         offsetX |= 0;
         offsetY |= 0;
         canvasWidth |= 0;
         canvasHeight |= 0;
-        var ctx: Context = new Context();
-        ctx.size(canvasWidth, canvasHeight);
-        ctx.asBitmap = true;
+        var ctx = rt?Sprite.drawtocanvCtx:new Context();
+		ctx.size(canvasWidth, canvasHeight);
+		if(rt){
+			ctx._targets=rt;
+		}else{
+			ctx.asBitmap=true;
+		}
         ctx._targets.start();
         ctx._targets.clear(0, 0, 0, 0);	// 否则没有地方调用clear
         RenderSprite.renders[_renderType]._fun(sprite, ctx, offsetX, offsetY);
         ctx.flush();
         ctx._targets.end();
-        ctx._targets.restore();
-        var rtex: Texture = new Texture(((<Texture2D>(ctx._targets as any))), Texture.INV_UV);
-        ctx.destroy(true);// 保留 _targets
-        return rtex;
+		ctx._targets.restore();
+		if(!rt){
+        	var rtex: Texture = new Texture(((<Texture2D>(ctx._targets as any))), Texture.INV_UV);
+        	ctx.destroy(true);// 保留 _targets
+			return rtex;
+		}
+		return rt;
     }
 
     /**
